requirements.txt
Django==1.9.7
Pillow
mysqlclient
python-dateutil
django-crispy-forms==1.6.0
